refactor: Address final pg_partman review comments
